requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.3
pandas==0.25.3
PyDSTool==0.90.3
pyparsing==2.4.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
scipy==1.3.1
six==1.13.0
bibtexparser=1.1.0
